+++ src/main/java/com/avrix/commands/Command.java
@@ -8,7 +8,6 @@ import zombie.core.raknet.UdpConnection;
  * Interface for implementing a new command.
  * Implementing classes must use the {@link CommandName} annotation to specify the command name.
  * They may also use the {@link CommandAccessLevel} annotation to specify the required access level for the command.
- * The {@link CommandChatReturn} annotation can be used to provide text that will be displayed in chat when the command is invoked.
  * The {@link CommandExecutionScope} annotation should be used to define where the command is available (e.g., in chat, in console, or both).
  * The {@link CommandDescription} annotation can be used to add a description to the command, providing additional information about its functionality.
  */
@@ -18,8 +17,9 @@ public abstract class Command {
      *
      * @param playerConnection {@link UdpConnection}, if called from the console, the connection will return as {@code null}
      * @param args             arguments of the received command
+     * @return message text that will be displayed to the user (or console) when the command is executed
      */
-    public abstract void onInvoke(UdpConnection playerConnection, String[] args);
+    public abstract String onInvoke(UdpConnection playerConnection, String[] args);
 
     /**
      * Retrieves the command name from the {@link CommandName} annotation.
@@ -48,16 +48,6 @@ public abstract class Command {
      */
     public AccessLevel getAccessLevel() {
         CommandAccessLevel annotation = this.getClass().getAnnotation(CommandAccessLevel.class);
-        return (annotation != null) ? annotation.value() : null;
-    }
-
-    /**
-     * Retrieves the chat return message from the {@link CommandChatReturn} annotation.
-     *
-     * @return the chat return message, or {@code null} if the annotation is not present
-     */
-    public String getChatReturnText() {
-        CommandChatReturn annotation = this.getClass().getAnnotation(CommandChatReturn.class);
         return (annotation != null) ? annotation.value() : null;
     }